# nestjs-server/src/swagger/api-metadata.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/users": {
      "get": {
        "description": "유저 목록 조회",
        "operationId": "AdminController_getUsers",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "1 페이지 내 사용자 수",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "description": "현재 페이지",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "필터링 조건 (JSON 문자열)\n예: {\"role\": 100}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponse"
                    },
                    {
                      "properties": {
                        "timestamp": {
                          "type": "string"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ResGetUsers"
                        },
                        "error": {
                          "type": "object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/update/role": {
      "put": {
        "description": "유저 역할 업데이트",
        "operationId": "AdminController_updateUserRole",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReqAdminUpdateRole"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponse"
                    },
                    {
                      "properties": {
                        "timestamp": {
                          "type": "string"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ResUser"
                        },
                        "error": {
                          "type": "object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/admin/db/page": {
      "get": {
        "description": "컬렉션 데이터 페이지네이션 및 필터 조회",
        "operationId": "AdminController_getDBDataWithFilter",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponse"
                    },
                    {
                      "properties": {
                        "timestamp": {
                          "type": "string"
                        },
                        "data": {
                          "$ref": "#/components/schemas/ResDBData"
                        },
                        "error": {
                          "type": "object"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "Admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ReqAdminUpdateRole": {
        "type": "object",
        "properties": {
          "useridx": {
            "type": "number"
          },
          "role": {
            "enum": [
              100,
              1,
              0
            ],
            "x-enum-varnames": [
              "ADMIN",
              "USER",
              "GUEST"
            ],
            "type": "number"
          }
        },
        "required": [
          "useridx",
          "role"
        ]
      },
...
      "CommonResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string"
          },
          "data": {
            "type": "object"
          },
          "error": {
            "type": "object"
          }
        },
        "required": [
          "timestamp"
        ]
      },
...
      "ResUser": {
        "type": "object",
        "properties": {
          "useridx": {
            "type": "number"
          },
          "nickname": {
            "type": "string"
          },
          "role": {
            "type": "number",
            "enum": [
              100,
              1,
              0
            ],
            "x-enum-varnames": [
              "ADMIN",
              "USER",
              "GUEST"
            ]
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "useridx",
          "nickname",
          "role",
          "created_at"
        ]
      },
      "ResGetUsers": {
        "type": "object",
        "properties": {
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResUser"
            }
          }
        },
        "required": [
          "users"
        ]
      },
...
      "ResDBData": {
        "type": "object",
        "properties": {
          "data": {
            "description": "컬렉션 데이터",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "total": {
            "type": "number",
            "description": "전체 데이터 수"
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
...
    }
  }
}